Cleanup: Refactor ChatRoom component to utilize groqService and handleChatStream
- Added groqService for creating conversations.
- Integrated handleChatStream service for managing chat streaming.
