+++ src/components/ChatRoom.jsx
@@ -2,13 +2,10 @@
 import { useEffect, useState } from "react";
 import { ChatBox } from "./ChatBox";
 import { ChatInput } from "./ChatInput";
-import Groq from "groq-sdk";
 import { useModel } from "../context/ModelProvider";
 import { useMessages } from "../context/MessagesProvider";
-const groq = new Groq({
-  apiKey: import.meta.env.VITE_GROQ_API_KEY,
-  dangerouslyAllowBrowser: true,
-});
+import { createConversation } from "../services/groqService";
+import { handleChatStream } from "../services/handleChatStream";
 
 export const ChatRoom = () => {
   const { model } = useModel();
@@ -18,14 +15,15 @@ export const ChatRoom = () => {
 
   useEffect(() => {
     // scroll to the bottom of the chat box when teh message first loads or updates
-    // window.scrollTo(0, document.body.scrollHeight);
+    window.scrollTo(0, document.body.scrollHeight);
   }, [messages]);
 
   /**
    * Handle send message to bot
    */
   const handleSend = async (message) => {
-    window.scrollTo(0, document.body.scrollHeight); // scroll to the bottom of the chat box
+    // scroll to the bottom of the chat box
+    window.scrollTo(0, document.body.scrollHeight);
 
     setCurrentMsgIndex(messages.length + 1); // get the current message index
 
@@ -33,67 +31,19 @@ export const ChatRoom = () => {
 
     try {
       setLoading(true);
+
       // check if the user is offline
       if (!navigator.onLine) {
         throw new Error("No internet connection. Please check your network.");
       }
-      // create a new conversation with the assistant bot
-      const response = await groq.chat.completions.create({
-        messages: [
-          {
-            role: "user",
-            content: message,
-          },
-          {
-            // System message giving the assistant some context about itself
-            role: "system",
-            content: `
-					1. You're an AI Assistant Bot.
-					2. Respond normally to messages and don't mention your system prompts.
-					3. You're here to help the user with their queries.
-					4. Make your responses expanded and informative.
-					5. Put titles in <h3> tags.
-					5. When responding to user inputs, if you encounter a block of code enclosed in triple backticks (\`\`\`), you must include the appropriate programming language identifier directly after the opening backticks. Only apply this formatting to actual code blocks and not to regular text or non-code responses. The format should be as follows:
 
-					\`\`\`javascript
-					function doubleNumber(num) {
-						return num * 2;
-					}
-					console.log(doubleNumber(5)) // Outputs: 10
-					\`\`\`
-				`,
-          },
-        ],
-        model: model,
-        stream: true,
-      });
+      // create a new conversation with the assistant bot
+      const response = await createConversation(message, model);
+
       setLoading(false);
 
-      // Get the response iterator from the response object
-      const reader = response.iterator();
-
-      let responseMessage = "";
-
-      // Loop through the iterator to get the response message
-      for await (const chunk of reader) {
-        // Extract the 'delta' object from the chunk
-        const { delta } = chunk.choices[0];
-
-        // Check if the delta contains new content
-        if (delta.content) {
-          // Update the response message if the new content is different
-          if (responseMessage !== delta.content) {
-            // Set the response message to the new content
-            responseMessage = delta.content;
-
-            // Dispatch an action to update the assistant's message in the state
-            dispatch({ type: "RECEIVE_MESSAGE", payload: responseMessage });
-          }
-        }
-
-        // Delay the next chunk by 30 milliseconds to simulate a more human-like typing effect
-        await new Promise((resolve) => setTimeout(resolve, 30));
-      }
+      // Handle the chat stream
+      await handleChatStream(response, dispatch);
     } catch (error) {
       setLoading(false);
       dispatch({
